audit static_cast

+++ src/changes/InsertRemoveChange.cpp
@@ -30,7 +30,7 @@ void InsertRemoveChange::insert_or_remove(bool should_insert) {
                                  parent_number);
   } else {
     chords_model_pointer->remove(first_child_number,
-                                 static_cast<int>(json_children.size()),
+                                 json_children.size(),
                                  parent_number);
   }
 }
