+++ src/models/ChordsModel.hpp
@@ -68,9 +68,9 @@ class ChordsModel : public QAbstractItemModel {
 
   // direct methods: generally take SongIndexes or parent_numbers and are not
   // undoable
-  void insert(int first_child_number, const nlohmann::json &json_children,
+  void insert(size_t first_child_number, const nlohmann::json &json_children,
               int parent_number);
-  void remove(int first_child_number, int number_of_children,
+  void remove(size_t first_child_number, size_t number_of_children,
               int parent_number);
 
   void set_cell(const SongIndex &index, const QVariant &new_value);
